removed godly()
This is a good example of immature code that really has no business in a respectable code base. There 
is a huge difference between providing comprehensive security access mechanisms to prevent abuse and 
hardcoded hacks that tell off someone and then kill them if they attempt to use a command on a certain 
character.